requirements.txt
# Agentic
anthropic>=0.86.0
pydantic>=2.12.0
# Python Utils
python-dotenv>=1.2.0
requests>=2.31.0
PyYAML>=6.0.0
# Article Collection
newsapi>=0.1.0
serpapi>=1.0.0
# Database
sqlalchemy>=2.0.0
cryptography>=42.0.0
# HTML Parsing
beautifulsoup4>=4.12.0
# Fuzzy Matching
rapidfuzz>=3.0.0
# Approval Server
flask>=3.0.0
apscheduler>=3.10.0
itsdangerous>=2.1.0
bleach>=6.0.0
flask-wtf>=1.2.0
flask-limiter>=3.5.0
authlib>=1.3.0
# LLM Providers
anthropic>=0.86.0
google-genai>=1.0.0
# Observability
langfuse>=4.0.0